+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 * Add option to disable automatic compilation in power save mode
 * Convert automatic compilation settings to a combobox
+* Add inspection to check for LaTeX package updates
 * Add checkboxes to graphic insertion wizard for relative width or height
 
 ### Fixed